safari-534.53-branch/LayoutTests/platform/chromium-win/fast/dom/getBoundingClientRect-expected.txt

Client bounding rect for #1
testRect.top 44
testRect.left 8
testRect.width 300
testRect.height 100
testRect.bottom 144
testRect.right 308

FAIL testRect.top should be 42. Was 44.
PASS testRect.left is 8
PASS testRect.width is 300
PASS testRect.height is 100
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #2
testRect.top 196
testRect.left 8
testRect.width 320
testRect.height 120
testRect.bottom 316
testRect.right 328

FAIL testRect.top should be 192. Was 196.
PASS testRect.left is 8
PASS testRect.width is 320
PASS testRect.height is 120
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #3
testRect.top 368
testRect.left 18
testRect.width 300
testRect.height 100
testRect.bottom 468
testRect.right 318

FAIL testRect.top should be 362. Was 368.
PASS testRect.left is 18
PASS testRect.width is 300
PASS testRect.height is 100
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #4
testRect.top 499
testRect.left 37
testRect.width 142
testRect.height 142
testRect.bottom 641
testRect.right 179

FAIL testRect.top should be 491. Was 499.
PASS testRect.left is 37
PASS testRect.width is 142
PASS testRect.height is 142
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #5
testRect.top 672
testRect.left 8
testRect.width 300
testRect.height 260
testRect.bottom 932
testRect.right 308

FAIL testRect.top should be 662. Was 672.
PASS testRect.left is 8
PASS testRect.width is 300
FAIL testRect.height should be 234. Was 260.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #6
Known failure. Skipping.

Client bounding rect for #7
testRect.top 1296
testRect.left 8
testRect.width 295
testRect.height 199
testRect.bottom 1495
testRect.right 303

FAIL testRect.top should be 1230. Was 1296.
PASS testRect.left is 8
FAIL testRect.width should be 299. Was 295.
FAIL testRect.height should be 180. Was 199.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #8
testRect.top 1548
testRect.left 8
testRect.width 300
testRect.height 50
testRect.bottom 1598
testRect.right 308

FAIL testRect.top should be 1460. Was 1548.
PASS testRect.left is 8
PASS testRect.width is 300
FAIL testRect.height should be 46. Was 50.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #9
testRect.top 1650
testRect.left 8
testRect.width 300
testRect.height 70
testRect.bottom 1720
testRect.right 308

FAIL testRect.top should be 1556. Was 1650.
PASS testRect.left is 8
PASS testRect.width is 300
FAIL testRect.height should be 64. Was 70.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #10
testRect.top 1774
testRect.left 8
testRect.width 300
testRect.height 22
testRect.bottom 1796
testRect.right 308

FAIL testRect.top should be 1672. Was 1774.
PASS testRect.left is 8
PASS testRect.width is 300
FAIL testRect.height should be 20. Was 22.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #11
testRect.top 1876
testRect.left 10
testRect.width 147
testRect.height 22
testRect.bottom 1898
testRect.right 157

FAIL testRect.top should be 1768. Was 1876.
PASS testRect.left is 10
PASS testRect.width is 147
FAIL testRect.height should be 20. Was 22.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #12
testRect.top 1976
testRect.left 8
testRect.width 100
testRect.height 100
testRect.bottom 2076
testRect.right 108

FAIL testRect.top should be 1862. Was 1976.
PASS testRect.left is 8
PASS testRect.width is 100
PASS testRect.height is 100
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #13
testRect.top 2213
testRect.left 8
testRect.width 175
testRect.height 19
testRect.bottom 2232
testRect.right 183

FAIL testRect.top should be 2098. Was 2213.
PASS testRect.left is 8
FAIL testRect.width should be 180. Was 175.
FAIL testRect.height should be 18. Was 19.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #14
testRect.top 2285
testRect.left 8
testRect.width 300
testRect.height 239
testRect.bottom 2524
testRect.right 308

FAIL testRect.top should be 2166. Was 2285.
PASS testRect.left is 8
FAIL testRect.width should be 200. Was 300.
FAIL testRect.height should be 236. Was 239.
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

Client bounding rect for #15
testRect.top 2577
testRect.left 8
testRect.width 0
testRect.height 0
testRect.bottom 2577
testRect.right 8

FAIL testRect.top should be 2452. Was 2577.
PASS testRect.left is 8
PASS testRect.width is 0
PASS testRect.height is 0
PASS testRect.bottom is testRect.top + testRect.height
PASS testRect.right is testRect.left + testRect.width

PASS successfullyParsed is true

TEST COMPLETE
